+++ src/gt4py/backend/dace_backend.py
@@ -24,9 +24,6 @@ import dace.data
 import numpy as np
 from dace.sdfg.utils import inline_sdfgs
 from dace.serialize import dumps
-from dace.transformation.dataflow import TrivialMapElimination
-from dace.transformation.helpers import nest_state_subgraph
-from dace.transformation.interstate import InlineTransients
 
 from eve import codegen
 from eve.codegen import MakoTemplate as as_mako
@@ -45,6 +42,11 @@ from gt4py.utils import shash
 from gtc import common, gtir
 from gtc.dace.nodes import StencilComputation
 from gtc.dace.oir_to_dace import OirSDFGBuilder
+from gtc.dace.transformations import (
+    InlineThreadLocalTransients,
+    eliminate_trivial_maps,
+    nest_sequential_map_scopes,
+)
 from gtc.dace.utils import array_dimensions, layout_maker_factory, replace_strides
 from gtc.gtir_to_oir import GTIRToOIR
 from gtc.passes.gtir_k_boundary import compute_k_boundary
@@ -174,110 +176,14 @@ def _pre_expand_trafos(gtir_pipeline: GtirPipeline, sdfg: dace.SDFG, layout_map)
     return sdfg
 
 
-def _eliminate_trivial_maps(sdfg: dace.SDFG):
-    applied = True
-    while applied:
-        applied = False
-        for map_entry, state in sdfg.all_nodes_recursive():
-            if isinstance(map_entry, dace.nodes.MapEntry):
-                if map_entry.map.schedule in {
-                    dace.ScheduleType.Sequential,
-                    dace.ScheduleType.CPU_Multicore,
-                }:
-                    # exclude maps with empty edges as workaround for a bug in TrivialMapElimination
-                    if any(
-                        edge.data.data is None
-                        for edge in state.in_edges(map_entry)
-                        + state.out_edges(state.exit_node(map_entry))
-                    ):
-                        continue
-                    try:
-                        TrivialMapElimination.apply_to(
-                            state.parent, map_entry=map_entry, verify=True
-                        )
-                        applied = True
-                        break
-                    except ValueError:
-                        continue
-
-
-def _tile_local_temporaries(sdfg: dace.SDFG):
-    changed = True
-    while changed:
-        changed = False
-        for map_entry, state in filter(
-            lambda n: isinstance(n[0], dace.nodes.MapEntry), sdfg.all_nodes_recursive()
-        ):
-            nsdfg: dace.SDFG = state.parent
-            scope_subgraph = state.scope_subgraph(
-                map_entry, include_entry=False, include_exit=False
-            )
-            if len(scope_subgraph) > 1 or not isinstance(
-                scope_subgraph.nodes()[0], dace.nodes.NestedSDFG
-            ):
-                continue
-            nsdfg_node = scope_subgraph.nodes()[0]
-            inline_transients = InlineTransients()
-            inline_transients.setup_match(
-                nsdfg,
-                nsdfg.sdfg_id,
-                nsdfg.node_id(state),
-                {InlineTransients.nsdfg: state.node_id(nsdfg_node)},
-                expr_index=0,
-            )
-            if inline_transients.can_be_applied(state, expr_index=0, sdfg=nsdfg):
-                changed = True
-                toremove = InlineTransients._candidates(nsdfg, state, nsdfg_node)
-                InlineTransients.apply_to(nsdfg, nsdfg=nsdfg_node)
-                for name in toremove:
-                    array: dace.data.Array = nsdfg_node.sdfg.arrays[name]
-                    shape = [dace.symbolic.overapproximate(s) for s in array.shape]
-                    strides = [1]
-                    total_size = shape[0]
-                    for s in reversed(shape[1:]):
-                        strides = [s * strides[0], *strides]
-                        total_size *= s
-                    array.shape = shape
-                    array.strides = strides
-                    array.total_size = total_size
-                    array.storage = dace.StorageType.CPU_ThreadLocal
-                    array.lifetime = dace.AllocationLifetime.Persistent
-
-
-def _nest_sequential_map_scopes(sdfg: dace.SDFG):
-    visited = set()
-
-    def _process_map(sdfg: dace.SDFG, state: dace.SDFGState, map_entry: dace.nodes.MapEntry):
-        for node in state.scope_subgraph(map_entry, include_entry=False, include_exit=False):
-            if node in visited:
-                continue
-            if isinstance(node, dace.nodes.NestedSDFG):
-                _nest_sequential_map_scopes(node.sdfg)
-            elif isinstance(node, dace.nodes.MapEntry):
-                _process_map(sdfg, state, node)
-            visited.add(node)
-        if map_entry.schedule == dace.ScheduleType.Sequential:
-            map_entrys = [map_entry]
-            for me in reversed(map_entrys):
-                subgraph = state.scope_subgraph(me, include_entry=False, include_exit=False)
-                nest_state_subgraph(sdfg, state, subgraph)
-
-    for state in sdfg.nodes():
-        for map_entry in filter(
-            lambda n: isinstance(n, dace.nodes.MapEntry) and n not in visited, state.nodes()
-        ):
-            _process_map(sdfg, state, map_entry)
-            visited.add(map_entry)
-
-
 def _post_expand_trafos(sdfg: dace.SDFG):
     # DaCe "standard" clean-up transformations
     sdfg.simplify(validate=False)
 
-    _eliminate_trivial_maps(sdfg)
+    eliminate_trivial_maps(sdfg)
 
-    # Only has effect if schedule is CPU_Multicore,
-    # setting node.collapse causes the omp parallel statement to include collapse(n)
+    # Control the `#pragma omp parallel` statements: Fully collapse parallel loops,
+    # but set 1D maps to be sequential. (Typical domains are too small to benefit from parallelism)
     for node, _ in filter(
         lambda n: isinstance(n[0], dace.nodes.MapEntry), sdfg.all_nodes_recursive()
     ):
@@ -285,9 +191,9 @@ def _post_expand_trafos(sdfg: dace.SDFG):
         if node.schedule == dace.ScheduleType.CPU_Multicore and len(node.range) <= 1:
             node.schedule = dace.ScheduleType.Sequential
 
-    _tile_local_temporaries(sdfg)
+    sdfg.apply_transformations_repeated(InlineThreadLocalTransients)
     sdfg.simplify()
-    _nest_sequential_map_scopes(sdfg)
+    nest_sequential_map_scopes(sdfg)
 
 
 def _sdfg_add_arrays_and_edges(
